test: update tests with refactored classes
